moved seed all method to alg instead of network
